+++ mobile/cmp/navigator/NavigatorModel.js
@@ -5,9 +5,9 @@
  * Copyright © 2021 Extremely Heavy Industries Inc.
  */
 import {HoistModel, RefreshMode, RenderMode, XH} from '@xh/hoist/core';
-import {action, bindable, observable, makeObservable} from '@xh/hoist/mobx';
+import {action, bindable, computed, observable, makeObservable} from '@xh/hoist/mobx';
 import {ensureNotEmpty, ensureUniqueBy, throwIf, warnIf} from '@xh/hoist/utils/js';
-import {find, isEqual, keys, merge} from 'lodash';
+import {isFinite, find, isEqual, keys, merge} from 'lodash';
 import {page} from './impl/Page';
 import {PageModel} from './PageModel';
 
@@ -28,11 +28,17 @@ export class NavigatorModel extends HoistModel {
     /** @member {boolean} */
     track;
 
+    /** @member {boolean} */
+    swipeToGoBack;
+
     /** @member {RenderMode} */
     renderMode;
 
     /** @member {RefreshMode} */
     refreshMode;
+
+    /** @package {number} */
+    @observable swipeProgress = null;
 
     _navigator = null;
     _callback = null;
@@ -48,11 +54,24 @@ export class NavigatorModel extends HoistModel {
         return stack[stack.length - 1];
     }
 
+    /** @type boolean */
+    @computed
+    get swipeStarted() {
+        return isFinite(this.swipeProgress);
+    }
+
+    /** @type boolean */
+    @computed
+    get swipeComplete() {
+        return this.swipeProgress === 1;
+    }
+
     /**
      * @param {Object[]} pages - configs for PageModels, representing all supported
      *      pages within this Navigator/App.
      * @param {boolean} [track] - True to enable activity tracking of page views (default false).
      *      Viewing of each page will be tracked with the `oncePerSession` flag, to avoid duplication.
+     * @param {boolean} [swipeToGoBack] - True to enable 'swipe to go back' functionality.
      * @param {RenderMode} [renderMode] - strategy for rendering pages. Can be set per-page
      *      via `PageModel.renderMode`. See enum for description of supported modes.
      * @param {RefreshMode} [refreshMode] - strategy for refreshing pages. Can be set per-page
@@ -61,6 +80,7 @@ export class NavigatorModel extends HoistModel {
     constructor({
         pages,
         track = false,
+        swipeToGoBack = true,
         renderMode = RenderMode.LAZY,
         refreshMode = RefreshMode.ON_SHOW_LAZY
     }) {
@@ -73,6 +93,7 @@ export class NavigatorModel extends HoistModel {
 
         this.pages = pages;
         this.track = track;
+        this.swipeToGoBack = swipeToGoBack;
         this.renderMode = renderMode;
         this.refreshMode = refreshMode;
 
@@ -212,7 +233,7 @@ export class NavigatorModel extends HoistModel {
         }
     }
 
-    renderPage(model, navigator) {
+    renderPage = (model, navigator) => {
         const {init, key} = model;
 
         // Note: We use the special 'init' object to obtain a reference to the
@@ -248,7 +269,7 @@ export class NavigatorModel extends HoistModel {
     }
 
     @action
-    onPageChange() {
+    onPageChange = () => {
         this.disableAppRefreshButton = this.activePage?.disableAppRefreshButton;
         this.doCallback();
     }
@@ -258,4 +279,60 @@ export class NavigatorModel extends HoistModel {
         this._callback = null;
     }
 
+    @action
+    onDragStart = (e) => {
+        this.swipeProgress = null;
+
+        // Determine if this gesture could be a potential navigation swipe.
+        if (!this.swipeToGoBack || this.stack.length < 2) return;
+
+        // Prevent swipes from the left edge of the screen, to not conflict
+        // with the native browser back gesture.
+        if (e.nativeEvent.gesture.startEvent.center.pageX < 20) return;
+
+        // Loop through the touch targets to ensure it is safe to swipe
+        for (let el = e.target; el && el !== document.body; el = el.parentNode) {
+            // Don't conflict with grid header reordering.
+            if (el.classList.contains('xh-grid-header')) {
+                return;
+            }
+            // Ensure any left-scrolling element in the target path takes priority over swipe navigation.
+            // We must check this at the start of the gesture, as scroll position changes throughout.
+            if (el.scrollWidth > el.offsetWidth && el.scrollLeft > 0) {
+                return;
+            }
+        }
+
+        this.swipeProgress = 0;
+        this.consumeEvent(e);
+    }
+
+    @action
+    onDrag = (e) => {
+        if (!this.swipeStarted) return;
+        const {direction, deltaX} = e.gesture;
+
+        // If the direction ever deviates, cancel the gesture
+        if (direction !== 'right') {
+            this.swipeProgress = null;
+            return;
+        }
+
+        // Set normalised progress based on distance dragged
+        this.swipeProgress = Math.clamp(deltaX / 150, 0, 1);
+        this.consumeEvent(e);
+    }
+
+    @action
+    onDragEnd = (e) => {
+        if (!this.swipeStarted) return;
+        if (this.swipeComplete) XH.popRoute();
+        this.swipeProgress = null;
+        this.consumeEvent(e);
+    }
+
+    consumeEvent(e) {
+        e.preventDefault();
+        e.stopPropagation();
+    }
 }